+++ worlds/pokemon_bw/data/text/funny_dialog.py
@@ -1,0 +1,75 @@
+from .. import TextData
+
+table: list[TextData] = [
+    TextData("BlastSlimey", "story", 67, 0, 41, "A friend of mine once tried to sneak[NextLine]"
+                                                "onto the cycling road in Kanto,[Scroll][NextLine]"
+                                                "without having a bicycle.[End][NextLine]"
+                                                "But as soon as they stepped out of[NextLine]"
+                                                "the gate, they were suddenly sitting[Scroll][NextLine]"
+                                                "on an unfamiliar bicycle![End][NextLine]"
+                                                "What a strange world we live in...[Terminate]"),
+    TextData("FlamelStones", "system", 236, 0, 547, "This level 100 EV Trained Fluffy Fluffer[NextLine]"
+                                                    "will Fluff you up, on Arceus, you better[NextLine]"
+                                                    "watch your back. #GrassGangRiseUp![Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 21, "N: [c_be09_#1_1]"
+                                                         "Ya know, someone wanted me to do[NextLine]"
+                                                         "something... that I cannot do.[c_be09_#1_0][End][Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 22, "[c_be09_#1_1]"
+                                                         "You wanna know what?[c_be09_#1_0][End][Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 23, "[c_be09_#1_1]"
+                                                         "They wanted me to...[c_be09_#1_0][End][Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 24, "[c_be09_#1_1]"
+                                                         "...recite the entire Bee Movie script.[c_be09_#1_0]"
+                                                         "[End][Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 25, "[c_be09_#1_1]"
+                                                         "Wait, you don't know why I[NextLine]"
+                                                         "can't do that?[c_be09_#1_0][End][Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 26, "[c_be09_#1_1]"
+                                                         "It's really simple, but people[NextLine]"
+                                                         "tend to forget it.[c_be09_#1_0][End][Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 27, "[c_be09_#1_1]"
+                                                         "If I were to recite copyrighted material,[NextLine]"
+                                                         "which requires that material to be[Scroll][NextLine]"
+                                                         "included in the apworld, this game would[Scroll][NextLine]"
+                                                         "get kicked out of the Archipelago Discord[Scroll][NextLine]"
+                                                         "Server in an instant.[c_be09_#1_0][End][Terminate]"),
+    TextData("Serterd/BlastSlimey", "story", 112, 0, 28, "[c_be09_#1_1]"
+                                                         "So a meta joke it is.[c_be09_#1_0][End][Terminate]"),
+    TextData("Serterd", "story", 351, 0, 19, "Nett hier.[End][NextLine]"
+                                             "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                             "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 358, 0, 14, "Nett hier.[End][NextLine]"
+                                             "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                             "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 363, 0, 5, "Nett hier.[End][NextLine]"
+                                            "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                            "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 377, 0, 20, "Nett hier.[End][NextLine]"
+                                             "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                             "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 387, 0, 18, "Nett hier.[End][NextLine]"
+                                             "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                             "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 400, 0, 2, "Nett hier.[End][NextLine]"
+                                            "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                            "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 406, 0, 13, "Nett hier.[End][NextLine]"
+                                             "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                             "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 415, 0, 3, "Nett hier.[End][NextLine]"
+                                            "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                            "Württemberg?[Terminate]"),
+    TextData("Serterd", "story", 247, 0, 17, "Nett hier.[End][NextLine]"
+                                             "Aber waren sie schon mal in Baden-[Scroll][NextLine]"
+                                             "Württemberg?[Terminate]"),
+    TextData("Kuchiri", "system", 236, 0, 384, "Oh, so you allowed yourself in the[NextLine]"
+                                               "yaml to obtain this pokémon. We[NextLine]"
+                                               "won't tell anyone ;)[Terminate]"),
+    TextData("ZoZoBab", "system", 236, 0, 549, "You caught a Zoey![NextLine]"
+                                               "I love you Odin![NextLine]"
+                                               "Viva la Spongegang![Terminate]"),
+]
+"""
+    TextData("AAA", "system", xxx, 0, xxx, "AAA"),
+    TextData("AAA", "story", xxx, 0, xxx, "AAA"),
+"""